+++ packages/react/src/SearchFilters.js
@@ -37,10 +37,14 @@ FiltersBox.displayName = 'FiltersBox'
 
 let BasicSearchFilters = ({ setMode, trees, children, BasicFilters }) => (
   <div>
-    <Flex style={{ alignItems: 'center' }}>
-      <h1>Filters</h1>
-      <ToggleFiltersButton onClick={() => setMode('resultsOnly')} />
-      <TreePauseButton children={children} />
+    <Flex alignItems="center" justifyContent="space-between">
+      <Flex alignItems="center">
+        <h1>Filters</h1>
+        <ToggleFiltersButton onClick={() => setMode('resultsOnly')} />
+      </Flex>
+      <div>
+        <TreePauseButton children={children} />
+      </div>
     </Flex>
     <LabelledList list={trees} Component={BasicFilters} />
     <LinkButton onClick={() => setMode('builder')} style={{ marginTop: 15 }}>