+++ packages/react/src/queryBuilder/QueryBuilder.js
@@ -1,12 +1,9 @@
 import React from 'react'
 import F from 'futil-js'
-import _ from 'lodash/fp'
-import { observer } from 'mobx-react'
 import DDContext from './DragDrop/DDContext'
 import Group from './Group'
 import styles from '../styles'
-import { withNode } from '../utils/hoc'
-import { withTheme } from '../utils/theme'
+import { contexturifyWithoutLoader } from '../utils/hoc'
 import { useLens } from '../utils/react'
 
 let { background } = styles
@@ -40,11 +37,6 @@ let QueryBuilder = ({
   )
 }
 
-export default DDContext(
-  _.flow(
-    observer,
-    withNode,
-    withTheme
-  )(QueryBuilder),
-  { allowEmptyNode: true }
-)
+export default DDContext(contexturifyWithoutLoader(QueryBuilder), {
+  allowEmptyNode: true,
+})